requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
asyncpg==0.30.0
pydantic-settings==2.7.1
